+++ docs/changelog.md
@@ -13,5 +13,5 @@
 
 ## 2022-07-22
 
-1. wsudo -U will create a process without privilege escalation. If it is currently running under the System account, it will fail to start the Store App with reduced privileges.（wsudo -U 会创建未提权的进程，如果当前运行在 System 账户下，降权启动 Store App 会失败。）
+1. wsudo -U will create a process without privilege escalation. If it is currently running under the System account, wsudo -U cannot start the Store App directly, but it can start the Store App in the launched shell.（在系统权限下，wsudo -U 无法直接启动 Store App，但是可以在启动的 shell 中启动 Store App。）
 2. wsudo add -B --basic as default (permission inheritance mode)（wsudo 添加 `-B` `--basic` 作为默认执行模式，即继承父进程权限，而 -U 则总是启动未提权的进程。）